Roztřídění kódu
Složka pro itemy + přesunutí třídy Medkit a dalších funkcí do samotného souboru
